remove duplicate, flaky tests
